Updated to imu v2.0.3 and eigen v2.0.0

See merge request bolderflight/software/mpu9250!22

+++ src/mpu9250/mpu9250.cc
@@ -486,7 +486,8 @@ bool Mpu9250::ReadRegisters(uint8_t reg, uint8_t count, uint8_t *data) {
     i2c_->beginTransmission(config_.dev);
     i2c_->write(reg);
     i2c_->endTransmission(false);
-    uint8_t bytes_rx = i2c_->requestFrom(config_.dev, count);
+    uint8_t bytes_rx =
+      i2c_->requestFrom(static_cast<uint8_t>(config_.dev), count);
     if (bytes_rx == count) {
       for (std::size_t i = 0; i < count; i++) {
         data[i] = i2c_->read();
